Add: Library Checker `Lowest Common Ancestor` 提出用バイナリーを追加
