+++ Cargo.toml
@@ -85,6 +85,24 @@ path = "src/bin/library_checker/bundled/bipartitematching.rs"
 required-features = ["bundled"]
 
 [[bin]]
+name = "lc-lca"
+path = "src/bin/library_checker/lca.rs"
+
+[[bin]]
+name = "lc-lca-bundled"
+path = "src/bin/library_checker/bundled/lca.rs"
+required-features = ["bundled"]
+
+[[bin]]
+name = "lc-lca-by-euler-tour"
+path = "src/bin/library_checker/lca_by_euler_tour.rs"
+
+[[bin]]
+name = "lc-lca-by-euler-tour-bundled"
+path = "src/bin/library_checker/bundled/lca_by_euler_tour.rs"
+required-features = ["bundled"]
+
+[[bin]]
 name = "lc-convolution-mod"
 path = "src/bin/library_checker/convolution_mod.rs"
 
@@ -298,15 +316,6 @@ path = "src/bin/atcoder/abc422_g_balls_and_boxes.rs"
 [[bin]]
 name = "ac-abc422-g-balls-and-boxes-bundled"
 path = "src/bin/atcoder/bundled/abc422_g_balls_and_boxes.rs"
-required-features = ["bundled"]
-
-[[bin]]
-name = "lc-lca"
-path = "src/bin/library_checker/lca.rs"
-
-[[bin]]
-name = "lc-lca-bundled"
-path = "src/bin/library_checker/bundled/lca.rs"
 required-features = ["bundled"]
 
 [[bin]]
